services/web/requirements.txt
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
distlib==0.3.6
filelock==3.11.0
Flask==2.2.3
gunicorn==20.1.0
identify==2.5.22
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
msal==1.21.0
nodeenv==1.7.0
platformdirs==3.2.0
pre-commit==3.2.2
pycparser==2.21
PyJWT==2.6.0
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
urllib3==1.26.15
virtualenv==20.21.0
Werkzeug==2.2.3
